+++ unirest-mocks/src/main/java/kong/unirest/Routes.java
@@ -59,12 +59,12 @@ class Routes implements Assert {
 
     RawResponse exchange(HttpRequest request, Config config) {
         return getBestMatch(request)
-                .map(invocation -> invocation.getResponse(config))
+                .map(invocation -> invocation.getResponse(config, request))
                 .orElseGet(() -> {
                     Invocation i = new Invocation(this);
                     i.log(request);
                     invokes.add(i);
-                    return i.getResponse(config);
+                    return i.getResponse(config, request);
                 });
     }