+++ unirest-mocks/src/test/java/kong/tests/Base.java
@@ -25,7 +25,9 @@
 
 package kong.tests;
 
+import kong.unirest.JsonObjectMapper;
 import kong.unirest.MockClient;
+import kong.unirest.Unirest;
 import kong.unirest.UnirestAssertion;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -41,6 +43,7 @@ public class Base {
     @BeforeEach
     public void setUp() {
         client = MockClient.register();
+        Unirest.config().setObjectMapper(new JsonObjectMapper());
     }
 
     @AfterEach