+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 3.11.10 (pending)
+* issue #394 use the configured Object Mapper rather than always Json
 * internal pre-factorings to get ready for Unirest 4
 
 ## 3.11.09